+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 
 # NukeUI 0.x
 
+- Add `ImageView` that supports animated images and it now used by `LazyImageView`
+- Remove `LazyImageView` view for setting image, you can use the underlying `ImageView` directly
+
 ## NukeUI 0.5.0
 
 - Rework `LazyImage` to use `FetchImage` on all platforms